split off utils package and flatten dir struction
