+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.0.0",
   "private": "true",
   "workspaces": [
+    "packages/utils",
     "packages/core",
     "packages/hooks",
     "packages/plonk"
