+++ packages/core/src/__tests__/Frames.test.ts
@@ -1,4 +1,3 @@
-import now from '../../utils/now';
 import Frames from '../Frames';
 
 describe('Frames', () => {
