+++ packages/core/src/__tests__/Scale.test.ts
