+++ packages/core/src/__tests__/Sine.test.ts
