+++ packages/core/src/index.ts
@@ -1,4 +1,22 @@
-export * from './constants';
-export * from './math';
-export * from './timers';
-export * from './utils';
+export { default as Drunk, type DrunkOptions, type DrunkState } from './Drunk';
+export { default as Env, type EnvOptions, type EnvState } from './Env';
+export { default as Frames, type FramesOptions } from './Frames';
+export {
+  default as Metro,
+  type MetroOptions,
+  type TimerState,
+  type TimerCallback,
+} from './Metro';
+export { default as Rand, type RandOptions, type RandState } from './Rand';
+export {
+  default as Scale,
+  type ScaleOptions,
+  type ScaleRange,
+  type ScaleState,
+} from './Scale';
+export {
+  default as Sine,
+  type SineOptions,
+  type SineState,
+  SINE_PERIOD,
+} from './Sine';
