--- packages/core/src/math/flip.ts
@@ -1,5 +1,0 @@
-const flip = (n: number): number => {
-  return n * -1;
-};
-
-export default flip;
