--- packages/core/src/math/index.ts
@@ -1,18 +1,0 @@
-export { default as clamp } from './clamp';
-export { default as expo } from './expo';
-export { default as flip } from './flip';
-export { default as Drunk, type DrunkOptions, type DrunkState } from './Drunk';
-export { default as Env, type EnvOptions, type EnvState } from './Env';
-export { default as Rand, type RandOptions, type RandState } from './Rand';
-export {
-  default as Scale,
-  type ScaleOptions,
-  type ScaleRange,
-  type ScaleState,
-} from './Scale';
-export {
-  default as Sine,
-  type SineOptions,
-  type SineState,
-  SINE_PERIOD,
-} from './Sine';
