--- packages/core/src/timers/index.ts
@@ -1,7 +1,0 @@
-export { default as Frames, type FramesOptions } from './Frames';
-export {
-  default as Metro,
-  type MetroOptions,
-  type TimerState,
-  type TimerCallback,
-} from './Metro';
