--- packages/core/src/utils/index.ts
@@ -1,2 +1,0 @@
-export { default as now } from './now';
-export { default as ms, type FPS, TimeFormat } from './ms';
