--- packages/hooks/src/hooks/index.ts
@@ -1,2 +1,0 @@
-export { default as useFrames } from './useFrames';
-export { default as useMetro, type UseMetroOptions } from './useMetro';
