+++ packages/hooks/src/index.ts
@@ -1,1 +1,2 @@
-export * from './hooks';
+export { default as useFrames } from './useFrames';
+export { default as useMetro, type UseMetroOptions } from './useMetro';
