+++ packages/plonk/README.md
@@ -1,3 +1,5 @@
-# [plonk](https://github.com/prtcl/plonk)
+# plonk
 
 Tiny library that provides timers, envelopes, and random generators.
+
+Visit [GitHub](https://github.com/prtcl/plonk) for docs and info.
