+++ packages/utils/.npmignore
@@ -1,0 +1,3 @@
+src
+jest.config.js
+tsconfig.json
