+++ packages/utils/README.md
@@ -1,0 +1,3 @@
+# @prtcl/plonk-utils
+
+Utils and helpers for [plonk](https://github.com/prtcl/plonk)
