+++ packages/utils/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@prtcl/plonk-hooks",
+  "name": "@prtcl/plonk-utils",
   "version": "1.0.0-alpha.2",
-  "description": "React hook wrappers for plonk",
+  "description": "Utils and helpers for plonk",
   "author": "Cory O'Brien <cory@prtcl.cc>",
   "license": "MIT",
   "main": "dist/index.js",
@@ -15,22 +15,13 @@
   "scripts": {
     "build": "npx tsc",
     "clean": "rm -rf dist | true",
-    "test": "npx jest --env=jsdom"
-  },
-  "dependencies": {
-    "@prtcl/plonk-core": "^2.0.0-alpha.2"
-  },
-  "peerDependencies": {
-    "react": "17 - 18"
+    "test": "npx jest --env=node && npx jest --env=jsdom"
   },
   "devDependencies": {
-    "@testing-library/react": "^14.2.2",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.12.3",
-    "@types/react": "^18.2.74",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "react": "^18.2.0",
     "ts-jest": "^29.1.2",
     "typescript": "^5.4.3"
   }
