+++ packages/utils/src/flip.ts
@@ -1,0 +1,4 @@
+/** Just flips the sign of incoming numbers. */
+export default function flip(n: number): number {
+  return n * -1;
+}
